+++ index.ts
@@ -1,0 +1,3 @@
+export * from "./src/types";
+export * from "./src/useTextInputMask";
+export { TextInputMask as default, TextInputMask } from "./src/TextInputMask";