+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+### Version 3.1.6
+
+- Extracted logic into `useTextInputMask` for use with custom `TextInput`s
+
 ### Version 2.0.0
 
-* Updated the headers in `RNTextInputMask.m` file to work with pods integration
+- Updated the headers in `RNTextInputMask.m` file to work with pods integration